Ralph Round 10: Scribe logging — 🔴 order regression flagged + player_input guard retraction
Logging summary:
- Merged Keaton-r10 (integration test refactor: tick_fixed_systems exposed to tests) + Keyser-r10 (phase-guard audit)
- 🔴 ORDER REGRESSION: popup_feedback_system + energy_system moved pre-despawn (breaks design intent)
  - Original order: despawn → feedback → display → energy
  - New order: runner contains [feedback, energy] → despawn → display
  - Severity: HIGH; fix in flight (Keaton-r11)
  - Comment at game_loop.cpp:188 now misleading; must revert or document

- [R10 CORRECTION] player_input guard claim (Keyser-r8/r9) retracted as WRONG
  - Guards protect dispatcher callbacks (game_state_system.disp.update<>()) outside runner
  - Test test_entt_dispatcher_contract.cpp:290 fails when guards dropped
  - Original recommendation kept visible; annotation added

- Test count claim (−14) refuted: actual +2 (no consolidation occurred)

Round 10 artifacts:
- Keaton-r10 pattern: Order-dependent regression tests must exercise production tick
- Keyser-r10 pattern (REVERSAL): Trace dispatcher.update<>() before declaring guard redundant
- Module health: playing_systems_runner 🔴, fixed_tick_runner 🟢
- Pending: CI grep check still awaiting user approval (Keyser-r9)
