+++ app/game_loop.cpp
@@ -171,25 +171,8 @@ void game_loop_init(entt::registry& reg,
 
 // ── Run ─────────────────────────────────────────────────────────────────────
 
-void tick_fixed_systems(entt::registry& reg, float dt) {
-    // game_state_system runs FIRST and owns the authoritative GoEvent /
-    // ButtonPressEvent drain for this tick (calls disp.update<GoEvent>() and
-    // disp.update<ButtonPressEvent>() at its top).  All pre-tick enqueues from
-    // input_system and gesture_routing are delivered
-    // here to listeners in registration order (see wire_input_dispatcher).
-    // Systems later in this list that also call disp.update<T>() (e.g.,
-    // player_input_system) will find an empty queue and execute as no-ops.
-    game_state_system(reg, dt);
-    song_playback_system(reg, dt);
-    tick_playing_systems(reg, dt);
-    // Keep obstacle lifecycle systems contiguous while obstacle component pools
-    // are still hot in cache from scroll/collision/miss/scoring passes.
-    obstacle_despawn_system(reg, dt);
-    // Keep score-feedback chain contiguous (queue -> popup spawn -> popup state)
-    // while popup/score pools are warm.
-    popup_display_system(reg, dt);
-    particle_system(reg, dt);
-}
+// tick_fixed_systems is defined in systems/fixed_tick_runner.cpp (shapeshifter_lib)
+// so integration tests can call it without the full render/input graph.
 
 // One frame: input → fixed timestep → render → blit → audio.
 // Not in header — called by game_loop_run and platform_run_loop (Emscripten).
