+++ tests/test_collision_system.cpp
@@ -1,7 +1,6 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_floating_point.hpp>
 #include "test_helpers.h"
-#include "game_loop.h"
 
 TEST_CASE("collision: shape gate cleared with matching shape", "[collision]") {
     auto reg = make_registry();
